+++ src/pages/staking/[[...staking_page]].tsx
@@ -3,24 +3,23 @@ import { NextPage } from 'next';
 import { useRouter } from 'next/router';
 import ErrorPage from '@components/ErrorPage';
 import DashboardMenu from '@components/dashboard/DashboardMenu';
-import DashboardPage from '@components/dashboard/pages/DashboardPage';
+import DashboardPage from '@components/staking/pages/DashboardPage';
 import StakePositionsPage from '@components/dashboard/pages/StakePositionsPage';
-import AddStakePage from '@components/dashboard/pages/AddStakePage';
+import AddStakePage from '@components/staking/pages/AddStakePage';
 import UnlockVested from '@components/dashboard/pages/UnlockVested';
 import ClaimTokens from '@components/dashboard/pages/ClaimTokens';
 import TransactionHistory from '@components/dashboard/pages/TransactionHistoryPage';
 
-const Dashboard: NextPage = () => {
+const Staking: NextPage = () => {
   const router = useRouter()
-  const { dashboard_page } = router.query
-  const route = dashboard_page?.toString()
+  const { staking_page } = router.query
+  const route = staking_page?.toString()
+  console.log(route)
 
   const pageMapper: { [key: string]: React.ReactElement } = {
     "transaction-history": <TransactionHistory />,
-    "add-stake": <AddStakePage />,
-    "manage-stake": <StakePositionsPage />,
-    // "unlock-vested": <UnlockVested />,
-    // "claim-tokens": <ClaimTokens />
+    "add": <AddStakePage />,
+    "manage-stake": <StakePositionsPage />
   }
 
   return <DashboardMenu>
@@ -33,4 +32,4 @@ const Dashboard: NextPage = () => {
   </DashboardMenu>
 };
 
-export default Dashboard;
+export default Staking;
